Test the datatype compatibility.
Somehow the stupid SDK is treating long as int32 even though I am perfectly on a 64-bit platform. Need to change all long type into int64_T in the mex file

+++ functions/src/tabou_QAP_mex.cpp
@@ -7,7 +7,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
     return;
 }
 
-long unif(long low, long high)
+size_t unif(long low, long high)
 {
     return(low + long(double(high - low + 1) * rando()));
 }
